requirements.txt
pip==21.1.2
flask-swagger-ui==3.36.0
Flask==2.0.2
mysql-connector==2.2.9
PyJWT==2.3.0
